requirements.txt
albucore==0.0.12
albumentations==1.4.11
altair==5.3.0
annotated-types==0.7.0
attrs==23.2.0
black==24.4.2
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
contourpy==1.2.1
cycler==0.12.1
cython==3.0.10
distlib==0.3.8
easydict==1.13
eval-type-backport==0.2.0
exceptiongroup==1.2.1
filelock==3.15.4
flatbuffers==24.3.25
fonttools==4.53.1
fsspec==2024.6.1
gitdb==4.0.11
gitpython==3.1.43
humanfriendly==10.0
identify==2.5.36
idna==3.7
imageio==2.34.2
iniconfig==2.0.0
insightface==0.7.3
jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lazy-loader==0.4
loguru==0.7.2
markdown-it-py==3.0.0
markupsafe==2.1.5
matplotlib==3.9.1
mdurl==0.1.2
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.3
nodeenv==1.9.1
numpy==1.26.4
onnx==1.16.1
onnxruntime==1.18.1
opencv-python-headless==4.10.0.84
packaging==24.1
pandas==2.2.2
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.2
pluggy==1.5.0
pre-commit==3.7.1
prettytable==3.10.0
protobuf==5.27.2
pyarrow==16.1.0
pydantic==2.8.2
pydantic-core==2.20.1
pydeck==0.9.1
pygments==2.18.0
pyparsing==3.1.2
pytest==8.2.2
pytest-asyncio==0.23.7
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyyaml==6.0.1
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.19.0
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.14.0
setuptools==69.2.0
six==1.16.0
smmap==5.0.1
streamlit==1.36.0
sympy==1.13.0
tenacity==8.5.0
threadpoolctl==3.5.0
tifffile==2024.7.2
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torch==2.3.1
torchvision==0.18.1
tornado==6.4.1
tqdm==4.66.4
typing-extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
virtualenv==20.26.3
wcwidth==0.2.13
websockets==12.0
wheel==0.43.0

opencv-python~=4.10.0.84
torchmetrics~=1.4.1
huggingface-hub~=0.23.4
clip~=1.0

plotly~=5.24.0
